statemanager/common: extract state manager interface to common
